refactor: extract gradebook into app/gradebook/ (Django-style app)
Pilot feature for the per-app restructure — gradebook was already the
most self-contained blueprint (own model file, own service, own
templates/gradebook/ subfolder), making it the lowest-risk proof of the
migration recipe used for later phases:

- app/models/gradebook.py -> app/gradebook/models.py
- app/blueprints/gradebook.py -> app/gradebook/routes.py (blueprint now
  declares template_folder='templates' so render_template('gradebook/...')
  call sites are unchanged)
- app/services/gradebook_service.py -> app/gradebook/services.py
- app/templates/gradebook/ -> app/gradebook/templates/gradebook/
- tests/test_gradebook.py + test_gradebook_bugfixes.py ->
  app/gradebook/tests/, with gradebook-only fixtures moved to a local
  conftest.py and cross-feature fixtures (course, active_school, quiz,
  assignment, client, ...) shared via new app/conftest.py so pytest's
  conftest discovery covers every future app/<feature>/tests/ package.

app/models/__init__.py's eager-import aggregator now sources GradeCategory
et al. from app.gradebook.models, preserving both its import order (a real
SQLAlchemy mapper-configuration dependency) and the public `from app.models
import X` surface every other call site already uses. Cross-references from
quiz.py and several test files that imported app.models.gradebook /
app.services.gradebook_service directly are updated to the new paths.

Static CSS/JS stay under app/static/css/gradebook/ and app/static/js/gradebook/
(not moved into app/gradebook/static/) — see plan notes on why: Flask's
built-in `static` endpoint and the app's asset_hash() cache-busting helper
both read from the single app.static_folder, and moving the physical files
would require rewriting every url_for('static', ...) call site for no
organizational gain, since these assets are already namespaced by feature.

Verified: create_app() builds with all 224 routes intact, SQLAlchemy
metadata carries all 33 tables (incl. every gradebook table), and the
full pytest suite is unchanged from baseline (25 pre-existing failures,
99 passing — identical set, just relocated in the report).
