+++ tests/test_stage3_features.py
@@ -4,8 +4,8 @@ Tests for Stage 3 Features (GAP-6: Remedial, GAP-1: CTT Analysis)
 import pytest
 from app.core.extensions import db
 from app.models import User, UserRole, Course, Quiz, QuizStatus
-from app.models.gradebook import GradeCategory, GradeCategoryType, GradeItem, GradeEntry, LearningObjective
-from app.services.gradebook_service import needs_remedial, get_remedial_label
+from app.gradebook.models import GradeCategory, GradeCategoryType, GradeItem, GradeEntry, LearningObjective
+from app.gradebook.services import needs_remedial, get_remedial_label
 
 
 class TestGAP6Remedial:
